backend/requirements.txt
fastapi==0.97.0
numpy==1.26.4
uvicorn==0.22.0
flatland-rl==4.1.1